app/src/main/java/com/codesoom/assignment/application/UserCommandService.java
package com.codesoom.assignment.application;

import com.codesoom.assignment.domain.User;
import com.codesoom.assignment.domain.UserRepository;
import com.codesoom.assignment.dto.UserData;
import org.springframework.stereotype.Service;

import javax.transaction.Transactional;

@Service
@Transactional
public class UserCommandService implements UserCreator, UserDelete {
    private final UserRepository userRepository;

    public UserCommandService(UserRepository userRepository) {
        this.userRepository = userRepository;
    }

    @Override
    public User createUser(UserData userData) {
        return this.userRepository.save(userData);
    }

    @Override
    public void delete(Long id) {
        this.userRepository.deleteById(id);
    }
}